+++ src/modules/wearedevs-lib/index.ts
@@ -1,13 +1,10 @@
-require("./res/css")
 import "./settings"
+require("./res/css")
 export { SettingsForm, SettingsSection, SectionInputType } from "./settings"
 export { SitePopup, SiteNotification } from "./dialogs"
-export { SitePopupPreset, SitePopupWithPreset } from "./dialogs"
+export { SitePopupPreset, SitePopupWithPreset, SitePopupYesNoResponse } from "./dialogs"
 
-export {
-    extractThreadIdFromUrl,
-    extractUserInfoFromTag,
-} from "./utils/data-extractor"
+export { extractThreadIdFromUrl, extractUserInfoFromTag } from "./utils/data-extractor"
 
 export { LocalUser } from "./utils/local-user"
 export { LinkType, getLinkType } from "./utils/link-type"
