+++ src/scripts-v2/settings/constants.ts
@@ -1,0 +1,3 @@
+export const DB_THEME_MODIFICATIONS = "wrdplus_theme_mods"
+export const DB_ADVANCED = "wrdplus_advanced_settings"
+export const DB_FEATURES = "wrdplus_features"
