+++ src/scripts-v2/settings/index.ts
@@ -1,1 +1,3 @@
 import "./main"
+export * as easyLoad from "./easy-load"
+export * as DB from "./constants"
